[refactor] :recycle: adapt helpers to new types and services (#276)
